+++ JohnSL_Trigger_Rewritten_For_Trinket_usingInternalPullUps/JohnSL_Trigger_Rewritten_For_Trinket_usingInternalPullUps.ino
@@ -33,13 +33,13 @@ short fsrAnalogNum[] = { 1, 2, 3 };     // Pins for each of the FSR analog input
 
 unsigned long lastLongTime[3];              // Last time in millis that we captured a long-term sample
 uint16_t longSamples[3][LONG_SIZE];         // Used to keep a long-term average
-uint8_t longIndex[3] = {0, 0, 0};           // Index of the last long-term sample
+uint16_t longIndex[3] = {0, 0, 0};           // Index of the last long-term sample
 uint16_t longAverage[3] = {0, 0, 0};
 
 uint16_t shortSamples[3][SHORT_SIZE];       // Used to create an average of the most recent samples
-uint8_t averageIndex[3] = {0, 0, 0};
+uint16_t averageIndex[3] = {0, 0, 0};
 
-uint8_t pin=0;                              // Defining a global pin variable that is used in the code.
+uint16_t pin=0;                              // Defining a global pin variable that is used in the code.
 
 void SetOutput(short fsr, bool state)
 {
@@ -48,7 +48,7 @@ void SetOutput(short fsr, bool state)
     triggered[fsr] = state;
     
     bool any = false;
-    for (uint8_t fsr = 0; fsr < 3; fsr++)
+    for (uint16_t fsr = 0; fsr < 3; fsr++)
     {
         any |= triggered[fsr];
     }
@@ -59,16 +59,16 @@ void SetOutput(short fsr, bool state)
 
 void InitValues()
 {
-    for (uint8_t fsr = 0; fsr < 3; fsr++)
+    for (uint16_t fsr = 0; fsr < 3; fsr++)
     {
-        for (uint8_t i = 0; i < SHORT_SIZE; i++)
+        for (uint16_t i = 0; i < SHORT_SIZE; i++)
             shortSamples[fsr][i] = 0;
 
-        for (uint8_t i = 0; i < LONG_SIZE; i++)
+        for (uint16_t i = 0; i < LONG_SIZE; i++)
             longSamples[fsr][i] = 0;
     }
 
-    for (uint8_t fsr = 0; fsr < 3; fsr++)
+    for (uint16_t fsr = 0; fsr < 3; fsr++)
         lastLongTime[fsr] = millis();
 }
 
@@ -79,7 +79,7 @@ void setup()
 {
     InitValues();
 
-    for (uint8_t fsr = 0; fsr < 3; fsr++)
+    for (uint16_t fsr = 0; fsr < 3; fsr++)
     {
         pin = fsrPins[fsr];
         pinMode(pin, INPUT);
@@ -145,8 +145,8 @@ void CalculateThreshold(short fsr)
 
 void loop()
 {
-//    uint8_t fsr = 8;
-    for (uint8_t fsr = 0; fsr < 3; fsr++)
+//    uint16_t fsr = 8;
+    for (uint16_t fsr = 0; fsr < 3; fsr++)
     {
         int value = analogRead(fsrAnalogNum[fsr]);
